feat: add new structural changes
Pay attention: not finished yet! Don't touch!
