--- src/main/java/tech/gruppone/stalker/server/model/GeographicalPoint.java
@@ -1,14 +1,0 @@
-package tech.gruppone.stalker.server.model;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-public class GeographicalPoint {
-
-  double latitude;
-
-  double longitude;
-
-}
