--- src/main/java/tech/gruppone/stalker/server/model/UnauthenticatedUser.java
@@ -1,11 +1,0 @@
-package tech.gruppone.stalker.server.model;
-
-import lombok.Data;
-import lombok.NonNull;
-
-@Data
-public class UnauthenticatedUser {
-
-  @NonNull String email;
-  @NonNull String password;
-}
