--- src/main/java/tech/gruppone/stalker/server/model/api/GetUsersResponse.java
@@ -1,13 +1,0 @@
-package tech.gruppone.stalker.server.model.api;
-
-import java.util.List;
-
-import lombok.Value;
-import tech.gruppone.stalker.server.model.User;
-
-@Value
-public class GetUsersResponse {
-  // TODO write me
-
-  List<User> users;
-}
