+++ src/main/java/tech/gruppone/stalker/server/models/api/AppConnection.java
@@ -1,0 +1,20 @@
+package tech.gruppone.stalker.server.models.api;
+
+import java.sql.Timestamp;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.NonNull;
+
+@Data
+@NoArgsConstructor
+public class AppConnection {
+
+  @NonNull
+  Long organizationId;
+
+  @NonNull
+  Long userId;
+
+  Timestamp creationDateTime;
+
+}
