+++ src/main/java/tech/gruppone/stalker/server/models/api/CurrentPlaces.java
@@ -1,0 +1,20 @@
+package tech.gruppone.stalker.server.models.api;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.NonNull;
+
+@Data
+@NoArgsConstructor
+public class CurrentPlaces {
+
+  @NonNull
+  Double timeInside;
+
+  @NonNull
+  Long organizationId;
+
+  @NonNull
+  Long placeId;
+
+}
