+++ src/main/java/tech/gruppone/stalker/server/models/api/LdapConfiguration.java
@@ -1,0 +1,20 @@
+package tech.gruppone.stalker.server.models.api;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.NonNull;
+
+@Data
+@NoArgsConstructor
+public class LdapConfiguration {
+
+  @NonNull
+  String host;
+
+  @NonNull
+  String username;
+
+  @NonNull
+  String password;
+
+}
