+++ src/main/java/tech/gruppone/stalker/server/models/api/Organization.java
@@ -1,0 +1,18 @@
+package tech.gruppone.stalker.server.models.api;
+
+import org.springframework.data.annotation.Id;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.NonNull;
+
+@Data
+@NoArgsConstructor
+public class Organization {
+
+  @Id
+  Long id;
+
+  @NonNull
+  OrganizationData data;
+
+}
