+++ src/main/java/tech/gruppone/stalker/server/models/api/OrganizationData.java
@@ -1,0 +1,33 @@
+package tech.gruppone.stalker.server.models.api;
+
+import java.sql.Timestamp;
+import java.util.ArrayList;
+import java.util.List;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.NonNull;
+import lombok.Singular;
+
+@Data
+@NoArgsConstructor
+public class OrganizationData {
+
+  @NonNull
+  String name;
+
+  @NonNull
+  String description;
+
+  LdapConfiguration ldapConfiguration;
+
+  @Singular
+  List<Place> places = new ArrayList<Place>();
+
+  @NonNull
+  Boolean isPrivate;
+
+  Timestamp creationDateTime;
+
+  Timestamp lastChangeDateTime;
+
+}
