+++ src/main/java/tech/gruppone/stalker/server/models/api/PlaceInfo.java
@@ -1,0 +1,23 @@
+package tech.gruppone.stalker.server.models.api;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.NonNull;
+
+@Data
+@NoArgsConstructor
+public class PlaceInfo {
+
+  @NonNull
+  String address;
+
+  @NonNull
+  String city;
+
+  @NonNull
+  String zipcode;
+
+  @NonNull
+  String state;
+
+}
