+++ src/main/java/tech/gruppone/stalker/server/models/api/Places.java
@@ -1,0 +1,17 @@
+package tech.gruppone.stalker.server.models.api;
+
+import lombok.NonNull;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+public class Places {
+
+  @NonNull
+  Long placeId;
+
+  @NonNull
+  Integer usersInside;
+
+}
