+++ src/main/java/tech/gruppone/stalker/server/models/api/request/PostUsersRequest.java
@@ -1,0 +1,19 @@
+package tech.gruppone.stalker.server.models.api.request;
+
+import lombok.Data;
+import lombok.NoArgsConstructor;
+import lombok.NonNull;
+import tech.gruppone.stalker.server.models.api.LoginData;
+import tech.gruppone.stalker.server.models.api.UserData;
+
+@Data
+@NoArgsConstructor
+public class PostUsersRequest {
+
+  @NonNull
+  LoginData loginData;
+
+  @NonNull
+  UserData userData;
+
+}
