+++ src/main/java/tech/gruppone/stalker/server/models/api/response/GetUsersResponse.java
@@ -1,0 +1,13 @@
+package tech.gruppone.stalker.server.models.api.response;
+
+import java.util.List;
+
+import lombok.Value;
+import tech.gruppone.stalker.server.models.api.User;
+
+@Value
+public class GetUsersResponse {
+  // TODO write me
+
+  List<User> users;
+}
